Reto_Flask-MySQLdb/Reto_Flask-MySQLdb/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
click==8.0.1
colorama==0.4.4
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap4==4.0.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
typing-extensions==3.10.0.0
visitor==0.1.3
Werkzeug==2.0.1
zipp==3.5.0
backports.entry-points-selectable==1.1.0
click==8.0.1
colorama==0.4.4
distlib==0.3.2
dominate==2.6.0
filelock==3.0.12
Flask==2.0.1
Flask-Bootstrap4==4.0.2
Flask-MySQLdb==0.2.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
platformdirs==2.0.2
pygame==2.0.1
six==1.16.0
typing-extensions==3.10.0.0
virtualenv==20.6.0
visitor==0.1.3
Werkzeug==2.0.1
zipp==3.5.0
